+++ public/device-api.ts
@@ -81,7 +81,13 @@ export enum ADITimePointInfoFlags {
    kTPInfoDefault = 0 | 0,
    kDeviceSyncRoundTrip = 1 | 0,
    kDeviceSyncUSBFrameTimes = 2 | 0,
-   kDeviceSynchUSBLocked = 4 | 0
+   kDeviceSynchUSBLocked = 4 | 0,
+   kDeviceSynchUSBStartOnSpecifiedFrame = 8 | 0,
+   kDeviceSynchUSBFullSuppport = kDeviceSyncRoundTrip |
+      kDeviceSyncUSBFrameTimes |
+      kDeviceSynchUSBLocked |
+      kDeviceSynchUSBStartOnSpecifiedFrame |
+      0
 }
 
 export class TimePointInfo {
@@ -512,7 +518,7 @@ export interface ProxyDeviceSys {
    ): void;
 }
 
-//Data stream configuration information Quark looks for in the 
+//Data stream configuration information Quark looks for in the
 //configuration parameter passed to ProxyDeviceSys.setupDataInStream()
 export interface IDeviceStreamConfiguration {
    dataFormat: BlockDataFormat;
@@ -661,7 +667,7 @@ export enum SysStreamEventType {
 
 export interface IDeviceProxyAPI {
    /**
-    * Invokes an arbitrary function on the JS ProxyDevice on the devices thread. 
+    * Invokes an arbitrary function on the JS ProxyDevice on the devices thread.
     * This is a general mechanism whereby user actions done
     * on the main (UI) thread can be applied down to the device script.
     *
@@ -671,7 +677,7 @@ export interface IDeviceProxyAPI {
     */
 
    /**
-    * Invokes an arbitrary function on the JS ProxyDevice on the devices thread. 
+    * Invokes an arbitrary function on the JS ProxyDevice on the devices thread.
     * This is a general mechanism whereby user actions done
     * on the main (UI) thread can be applied down to the device script.
     *
@@ -685,10 +691,7 @@ export interface IDeviceProxyAPI {
       functionArgJson: string,
       checkExistsOnly?: boolean
    ): Promise<Record<string, any> | null>;
-
 }
-
-
 
 export interface IDeviceManagerApi {
    dispose(): void;
@@ -759,7 +762,7 @@ export interface IDeviceManagerApi {
    ): void;
 
    /**
-    * Deprecated! Use the simpler and more general IDeviceProxyAPI.callFunction() instead! 
+    * Deprecated! Use the simpler and more general IDeviceProxyAPI.callFunction() instead!
     * Invokes an arbitrary function somewhere within the device manager settings
     * on the devices thread. This is the general mechanism whereby user actions done
     * on the main (UI) thread are applied down to the device hardware.
@@ -798,7 +801,7 @@ export interface IDataSink {
    ): void;
    onPacket?(packetType: unknown, buffer: unknown): void;
    onError(err: Error): void;
-   inputToStream?: number[];  //mapping from device inputs to device output streams
+   inputToStream?: number[]; //mapping from device inputs to device output streams
 }
 
 //The JS part of the ProxyDevice called from Quark
@@ -825,7 +828,9 @@ export interface IProxyDevice {
    //Allocate StreamRingBuffers buffers
    prepareForSampling(bufferSizeInSecs: number): boolean;
 
-   startSampling(): boolean;
+   //Start the device sampling. If the device capabilities include being able
+   //to start on a future USB frame, the frame number is passed in.
+   startSampling(startOnUSBFrame?: number): boolean;
 
    onSamplingStarted(): void;
    onSamplingUpdate(): void;
